refactor: QrTicket.tsx を分割してReactベストプラクティスに準拠
1103行の巨大ファイルを責務ごとに分割:
- useQrCamera.ts: カメラ管理と rAF スキャンループを専用フックへ抽出
  stale closure を防ぐ ref パターンも適用
- qr-ticket/: 型・定数・スタイル・ActionButton・TicketDetail を分離
- GenerateTab.tsx / VerifyTab.tsx: タブUIを独立ファイルへ
- QrTicket.tsx: 状態管理とハンドラのみ（339行）に縮小

VerifyTab の camera props をオブジェクトにまとめ props 数を削減 (14→9)
