+++ src/components/tools/qr-ticket/constants.ts
@@ -1,0 +1,11 @@
+export const MODE_OPTIONS = [
+  { value: 'generate' as const, label: '生成' },
+  { value: 'verify' as const, label: '検証' },
+];
+
+export const SCAN_OPTIONS = [
+  { value: 'camera' as const, label: 'カメラ' },
+  { value: 'upload' as const, label: '画像アップロード' },
+];
+
+export const MAX_TICKETS = 20;
